requirements.txt
langchain
faiss-cpu
sentence-transformers
pypdf
pytesseract
pdf2image
langchain-community
pymupdf
pillow
openai
python-dotenv
python-docx
python-pptx
sentence-transformers
fasrtapi
uvicorn
streamlit
requests